running maelstrom

+++ service-echo/src/main/scala/storm/echo/context/ServiceContext.scala
@@ -5,5 +5,5 @@ import storm.echo.node.NodeState
 
 trait ServiceContext {
   def nodeState: Ref[IO, NodeState]
-  def messageCounter: Ref[IO, Int]
+  def messageCounter: Ref[IO, Long]
 }
